Add read filter for gatk CollectSVEvidence
Sometimes an aligner will generate a read mapping with zero aligned
bases. When gatk tries to process such a read in the CollectSVEvidence
command, it will throw an exception because it gives the read a negative
length. This commit adds a read filter when running CollectSVEvidence to
ignore such reads.