Prevent duplicate lines when subsetting RD matrices
In order to subset all of the RD matrices, the VisualizeCnvs workflow
first uses `bedtools merge` to merge the intervals of all CNVs to plot
and then uses `tabix` to retrieve them from the matries. By default,
intervals that are not adjacent or do not overlap, are not merged.
However, when two intervals are close to each other, they could overlap
the same interval in an RD matrix then using tabix will retrive that
shared interval twice causing problems downstream. To address this
issue, we can merge CNV intervals that are close and remove duplicate
lines in the subset RD matrix.