Add a workaround for when a VCF is missing END2
The RescueMobileElementDeletions task in CleanVcfChromosome assumes that
BND variants always have an END2 value in the INFO column. However,
VCFs produced by older versions of the pipeline did not consistently
include an END2 value which can cause the task to fail. As a workaround
for those VCFs, we compute the END2 value from the variant start and
SVLEN. To maintain backwards compatiblity, we make the workaround opt-in
with a task input and keep the previous behavior as the default.