Add workflow to set genotypes to null
The NullGenotypes workflows is designed to a take a set of samples and a
set of contigs and modify the genotypes for the given samples on the
given contigs to be null. It is meant for use on chrX and chrY where the
calls on those chromosomes are suspect due to errors in the pedigree,
but the default contig list can be overridden for flexibility.