Only annotate samples in the VCF
It is possible that some input samples are missing from the input VCF
and that will cause bcftools to fail because it can't annotate those
samples. To avoid this error, we create a subset of samples that are
shared between the input VCF and input samples list and only annotate
those.