WIP: Fix MatchContigToVcf optional output
The `read_string` function will just error if its input argument does
not exist instead of returning a `String?`. Because the MatchContigToVcf
task relies on outputting optional values to indicate whether or not a
contig matched, using `read_string` by itself will not work. We can use
an optional input value as an output value to handle the case when there
were no matching contigs.