WIP: Use different input and output VCF names
In MatchContigToVcf, if the input VCF variable is used as the output VCF
variable, the output variable will never be null because the task will
just output whatever VCF was localized. A fix is to rename the VCF as
part of the command and use the new name as the output. Then the
localized VCF can be deleted and it will not be delocalized.