WIP: fix NameError
The `get_insertion_intersection` function in denovo_svs.py referenced a
variable that was not in scope. The referenced variable stores a
configuration option and is local to `main`. This commit adds the
referenced variable as an argument to `get_insertion_intersection` so
that it can be passed when the function is called in main.