WIP: pin yulab.utils to 0.1.9
yulab.utils version 0.2.0 added an R version dependency of >=4.2.0 which
makes it incompatible with the version of R in the gatk-sv denovo Docker
image which is 4.1.3. That make ggplotify incompatible because it
depends on yulab.utils. Installing yulab.utils 0.1.9 from CRAN archives
and then installing the latest ggplotify version (0.1.2) works.