denovo: remove reference to config
There is no longer a config file being used with `denovo_svs.py`, but
there was a reference to a config variable left in the script. The
reference was trying to access a "parents_SC" key, which is never
actually included in the config file so we just remove the reference.