denovo: add a column for `is_de_novo`
In `src/denovo/denovo_svs.py` there is a column called `is_de_novo` in
the child variant DataFrame that tracks which variants in the child are
predicted to be de novo as the script executes. Previously, the column
was initialized while filtering by allele frequency, but that step was
removed from the script and as a result the column is never initialized,
causing `KeyError`s. We add a step to initialize the column to `True`,
matching previous behavior.