denovo: read output file paths explicitly
The use of `glob` in `MergeBatchBedsToContigs` will cause the file paths
in the glob expansion to be delocalized to a glob subdirectory and the
paths of the delocalized files will not match the paths in the
output `contigs_beds` map. Creating a `Array[File]` output with paths
matching the paths in the map will keep the paths synced.