WIP: denovo: reorganize initial VCF filtering
There are different levels of filtering that can be applied to the input
VCF of the de novo workflow. It is useful to split them into separate
tasks and apply them from broadest to most specific both for efficiency
and clarity reasons. At the highest level, there are the cohort site
filters. De novo calling is not done for BND and mCNV sites so they can
be immediately removed. CPX and CTX sites are handled specially so they
can be separated from the other variants. At the next level down are the
proband SV filters. There are site and genotype filters that can be
applied to all probands to remove low confidence calls. At the next
level down, there are raw evidence filters where cohort-wide proband
genotypes are checked against the raw batch evidence. But this not yet
implemented.