WIP: denovo: do not use pipe with grep exit status
Using `set -o pipefail` in combination with a construct like
`command | grep -q` will sometimes fail. If grep finds the search
pattern and exists, it will close the pipe and `command` may exit with a
SIGPIPE which, due to the pipefail, will cause the whole pipeline to
inherit the SIGPIPE exit status and the search will appear to have
failed. There is some slightly complicated syntax that be used to get
around the issue, but in this use case of checking the BCF header, it is
simpler to just write the header to a file and use grep on the file.