denovo: pass OAuth token to every tabix invocation
The RunDeNovo task exports the Google Cloud OAuth token when it starts
so that tabix can access files on Google Cloud Storage, but if the task
takes too long, the token might expire and accessing remote files will
fail. As a patch, we always grab the latest token prior to invoking
tabix and include it in the environment of the subprocess.