denovo: set max AF for sites overlapped by GD regions
Passing all sites covered by genomic disorder regions through the VCF
pre-filter could allow some sites with unusually high allele frequencies
to slip through and greatly increase the computational burden in
downstream steps. Given that sites with cohort AF greater than 10% are
unlikely to be true de novo sites, it is a good idea to exclude them
early in the workflow. Because cohort AF depends on cohort size, the
threshold for max AF is left as a workflow input parameter.