WIP(denovo): swap truth and eval for SVConcordance
SVConcordance will only annotate each site in the eval VCF with a single
site in the truth VCF (if there is a match). If a sample has an SV at a
site in the eval VCF, but it does not have an SV at the matched site in
the truth VCF, it will appear as though the sample's genotype is
unsupported in the eval VCF. However, due to clustering and joint
genotyping, it is possible that there is a different site in the truth
VCF which is a worse match to the site in the eval VCF, but still
supports the sample's genotype in the eval VCF. If the eval and truth
VCF are swapped, each site in the eval VCF will be annotated with the
best matching site in the truth VCF and then each site in the truth VCF
can be checked for support at any matching site in the eval VCF. It may
be that a site in the eval VCF supports multiple site in the truth VCF,
but that would suggest clustering/genotyping issues.