WIP(denovo): build denovo Docker from Ubuntu
The denovo Docker image builds from the sv_pipeline Docker image even
though it does not need anything from that image resulting in an
unnecessarily large image size. Building from a clean Ubuntu image
allows the image to stay small. If enough dependencies that overlap with
ones in the sv_pipeline image are added in the future, the change can be
reverted.