WIP(denov): reduce scatter when processing offsprings
The large number of scatters due to the nested scatter adds a lot of
overhead to the workflow (processing by the execution engine, waiting
for a VM, etc). The overhead ends up being significant fraction of the
scattered tasks because the tasks usually have a short runtime. One
place to reduce the scatter is when grouping the offspring BCF by batch.
Instead of scattering separately over offspring, fathers and mothers,
for each batch, the grouping can be done in a single task per batch.