WIP(denovo): fix filtering genotypes against parents
In `src/denovo/filtergt.c`, the wrong pointer was being used to check
that a sample ID from the sample to parent association file was in the
BCF header so no entries were being added to the `parents` hash map.
That prevented any genotypes from being filtered. There was also a
trailing newline kept in the parental IDs which would have prevented
matching of IDs in the `genotypes` map.