WIP(denovo): fix offspring sites filter
A filename was used as part of a filtering expression in the
FilterOffspringSites task, but it was quoted which caused it to be
interpreted as a string and instead of comparing site IDs to the
contents of the file, the IDs were being compared to the filename with
an '@' preprended and no sites were being removed. The filename
expression has been unquoted to preform the intended comparison.