WIP(denovo): fix errors from empty output files
If FilterGenotypes outputs an empty file because all genotypes were
filtered, then MakeDeNovoCalls will fail because DuckDB will try to
uncompress an empty file. By modifying the de novo calling step to
counting lines that appear three times in the union of all the filtered
genotypes files (corresponding to an SV being present in the offspring
but not in either parents), empty files can be gracefully ignored.